Working round figure

+++ src/main/java/com/sdd/utils/DocxGenaratorUtil.java
@@ -48,7 +48,7 @@ public class DocxGenaratorUtil {
             XWPFParagraph paragraphtableRowOne11 = tableRowOne.addNewTableCell().addParagraph();
             boldText(paragraphtableRowOne11.createRun(), 10, filePathResponse.getType() + " (" + filePathResponse.getFinYear() + ") \n" + " ALLOCATION (In " + filePathResponse.getAmountType() + ")", true);
 
-            Float grandTotal = 0f;
+            double grandTotal = 0f;
             for (Map.Entry<String, List<ReportSubModel>> entry11 : hashMap.entrySet()) {
                 String key11 = entry11.getKey();
                 List<ReportSubModel> tabData11 = entry11.getValue();
@@ -56,7 +56,7 @@ public class DocxGenaratorUtil {
                 XWPFTableRow tableRow = table.createRow();
                 tableRow.getCell(0).setText(key11);
 
-                Float allAmountData = 0f;
+                double allAmountData = 0f;
                 for (Integer i = 0; i < tabData11.size(); i++) {
 
                     if (i == 0) {
@@ -81,8 +81,8 @@ public class DocxGenaratorUtil {
                         paragraph11.setAlignment(ParagraphAlignment.RIGHT);
                     }
 
-                    allAmountData = allAmountData + Float.parseFloat(tabData11.get(i).getAmount());
-                    grandTotal = grandTotal + Float.parseFloat(tabData11.get(i).getAmount());
+                    allAmountData = allAmountData + Double.parseDouble(tabData11.get(i).getAmount());
+                    grandTotal = grandTotal + Double.parseDouble(tabData11.get(i).getAmount());
 
                 }
 
@@ -160,7 +160,7 @@ public class DocxGenaratorUtil {
             boldText(paragraphtableRowOne1111.createRun(), 10, "", true);
 
 
-            Float grandTotal = 0f;
+            double grandTotal = 0f;
             for (Map.Entry<String, List<ReportSubModel>> entry11 : hashMap.entrySet()) {
                 String key11 = entry11.getKey();
                 List<ReportSubModel> tabData11 = entry11.getValue();
@@ -168,7 +168,7 @@ public class DocxGenaratorUtil {
 //                XWPFTableRow tableRow = table.createRow();
 //                tableRow.getCell(0).setText(key11);
 
-                Float allAmountData = 0f;
+                double allAmountData = 0f;
                 for (Integer i = 0; i < tabData11.size(); i++) {
 
                     XWPFTableRow tableRow11 = table.createRow();
@@ -181,7 +181,7 @@ public class DocxGenaratorUtil {
                     normalText(paragraph11.createRun(), 10, ConverterUtils.addDecimalPoint(tabData11.get(i).getAmount()), false);
 
                     allAmountData = allAmountData + Float.parseFloat(tabData11.get(i).getAmount());
-                    grandTotal = grandTotal + Float.parseFloat(tabData11.get(i).getAmount());
+                    grandTotal = grandTotal + Double.parseDouble(tabData11.get(i).getAmount());
 
                 }
 
@@ -246,7 +246,7 @@ public class DocxGenaratorUtil {
             boldText(paragraphtableRowOne11.createRun(), 10, filePathResponse.getType() + " (" + filePathResponse.getFinYear() + ") \n" + " ALLOCATION (In " + filePathResponse.getAmountType() + ")", true);
 
 
-            Float grandTotal = 0f;
+            double grandTotal = 0f;
             for (Map.Entry<String, List<ReportSubModel>> entry11 : hashMap.entrySet()) {
                 String key11 = entry11.getKey();
                 List<ReportSubModel> tabData11 = entry11.getValue();
@@ -266,7 +266,7 @@ public class DocxGenaratorUtil {
                 }
 
 
-                Float allAmountData = 0f;
+                double allAmountData = 0f;
                 for (Integer i = 0; i < tabData11.size(); i++) {
 
                     XWPFTableRow tableRow11 = table.createRow();
@@ -279,8 +279,8 @@ public class DocxGenaratorUtil {
                     normalText(paragraph11.createRun(), 10, ConverterUtils.addDecimalPoint(tabData11.get(i).getAmount()), false);
                     paragraph11.setAlignment(ParagraphAlignment.RIGHT);
 
-                    allAmountData = allAmountData + Float.parseFloat(tabData11.get(i).getAmount());
-                    grandTotal = grandTotal + Float.parseFloat(tabData11.get(i).getAmount());
+                    allAmountData = allAmountData + Double.parseDouble(tabData11.get(i).getAmount());
+                    grandTotal = grandTotal + Double.parseDouble(tabData11.get(i).getAmount());
 
                 }
 
@@ -339,7 +339,7 @@ public class DocxGenaratorUtil {
     }
 
 
-    public void createCdaMainReportDoc(HashMap<String, List<CDAReportResponse>> map, CDAReportSubResponse cadSubReport, String path, Float grandTotal, HashMap<String, String> coloumWiseAmount) throws Exception {
+    public void createCdaMainReportDoc(HashMap<String, List<CDAReportResponse>> map, CDAReportSubResponse cadSubReport, String path, double grandTotal, HashMap<String, String> coloumWiseAmount) throws Exception {
 
         try {
 
@@ -489,7 +489,7 @@ public class DocxGenaratorUtil {
     }
 
 
-    public void createReserveFundnReport(HashMap<String, List<CDAReportResponse>> map, CDAReportSubResponse cadSubReport, String path, Float grandTotal, Float allocationGrandTotal) throws Exception {
+    public void createReserveFundnReport(HashMap<String, List<CDAReportResponse>> map, CDAReportSubResponse cadSubReport, String path, double grandTotal, double allocationGrandTotal) throws Exception {
 
         try {
